Removed Jpa. Finished InMemoryPlayerRepository
I wanted more control with my methods so I removed Jpa extension from the PlayerRepository interface. Finished InMemoryPlayerRepository for now.

+++ Table-Football-Scorekeeper-API/src/main/java/com/football/Table_Football_Scorekeeper_API/Services/PlayerService.java
@@ -8,13 +8,13 @@ public interface PlayerService {
 
     Player addPlayer(Player player);
 
-    Optional<Player> getPlayer(int id);
+    Optional<Player> getPlayer(Long id);
 
     List<Player> getAllPlayers();
 
-    Optional<Player> updatePlayer(int id, String name);
+    Optional<Player> updatePlayer(Long id, String name);
 
-    boolean deletePlayer(int id);
+    boolean deletePlayer(Long id);
 
     // Custom methods
